refactor: move `crypto` module to `shared/`
We are moving everything out of the `common/` folder
since we do not need this anymore.
