+++ packages/core/src/codewhisperer/commands/startSecurityScan.ts
@@ -37,7 +37,7 @@ import path from 'path'
 import { ZipMetadata, ZipUtil } from '../util/zipUtil'
 import { debounce } from 'lodash'
 import { once } from '../../shared/utilities/functionUtils'
-import { randomUUID } from '../../common/crypto'
+import { randomUUID } from '../../shared/crypto'
 import { CodeAnalysisScope, ProjectSizeExceededErrorMessage } from '../models/constants'
 import { CodeScanJobFailedError, CreateCodeScanFailedError, SecurityScanError } from '../models/errors'
