+++ packages/core/src/shared/crypto.ts
@@ -21,7 +21,7 @@
  *
  * Once we do not need `crypto` anymore, we can get rid of the polyfill.
  */
-import { isWeb } from '../shared/extensionGlobals'
+import { isWeb } from './extensionGlobals'
 
 export function randomUUID(): `${string}-${string}-${string}-${string}-${string}` {
     if (isWeb()) {
