+++ packages/core/src/test/amazonqFeatureDev/session/sessionState.test.ts
@@ -21,7 +21,7 @@ import { MessagePublisher } from '../../../amazonq/messages/messagePublisher'
 import { FeatureDevClient } from '../../../amazonqFeatureDev/client/featureDev'
 import { ToolkitError } from '../../../shared/errors'
 import { PrepareRepoFailedError } from '../../../amazonqFeatureDev/errors'
-import * as crypto from '../../../common/crypto'
+import * as crypto from '../../../shared/crypto'
 import { TelemetryHelper } from '../../../amazonqFeatureDev/util/telemetryHelper'
 import { assertTelemetry, createTestWorkspaceFolder } from '../../testUtil'
 import { getFetchStubWithResponse } from '../..'
